seperated the methods which helps in download the stock data into a seperate util module
